requirements.txt
langchain
langchain-groq
pypdf
faiss-cpu
python-dotenv
sentence-transformers
streamlit
PyPDF2
langchain-community